backend/requirements.txt
alembic==1.18.4
aiofiles>=23.0.0
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.84.0
anyio==4.12.1
bcrypt>=4.1.0
certifi==2026.2.25
chardet>=5.0.0
cffi==2.0.0
click==8.3.1
cryptography==45.0.7
distro==1.9.0
docstring_parser==0.17.0
fastapi==0.133.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.13.0
Mako==1.3.10
MarkupSafe==3.0.3
lxml>=4.9.0
openai==2.24.0
openpyxl>=3.1.0
pandas>=2.0.0
psycopg2-binary==2.9.11
Pillow==11.3.0
pytz>=2024.1
pypdf==6.1.3
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
pycparser==2.23
PyJWT==2.10.1
python-dotenv==1.2.1
python-multipart==0.0.20
qrcode[pil]>=7.4.2
sniffio==1.3.1
SQLAlchemy==2.0.47
starlette==0.52.1
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.41.0
